backend/requirements.txt
# Core dependencies
Django>=5.2.7
djangorestframework>=3.16.1
psycopg2-binary>=2.9.11
python-dotenv>=1.2.1

# Authentication
djangorestframework-simplejwt>=5.3.1

# CORS
django-cors-headers>=4.9.0

# API Documentation
drf-spectacular>=0.28.0

# AI Integration
openai>=2.6.1

# HTTP Requests
requests>=2.31.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0

# Testing
pytest>=7.4.3
pytest-django>=4.7.0
factory-boy>=3.3.0
coverage>=7.3.2